Añadidos los tests de mis 2 primeras historias, además arreglos en el delete y en la internacionalización

+++ src/main/java/com/uniovi/sdi2324entrega181/services/UsersService.java
@@ -99,11 +99,12 @@ public class UsersService {
         return users;
     }
 
-    public void borrarPorId(List<Long> usuariosABorrar) {
+    public void borrarPorId(List<Long> usuariosABorrar, String correo) {
 
         for(Long id: usuariosABorrar){
-
-            usersRepository.deleteById(id);
+            User principal = getUserByEmail(correo);
+            if(principal.getId() != id)
+                usersRepository.deleteById(id);
         }
     }
 }
